fix: Skip service runtime operations in non-booted hosts
Cause: Both the actual role and the tests were previously assuming that
the system was booted with systemd and could run NetworkManager, detect
and change hardware. But this is not the case when running the role
during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not.

Result: The role now works during container builds.

Some of the hardware setup steps in tests run before role invocation,
so they can't use `__network_is_booted` yet. In these cases check
`ansible_connection == 'buildah'`. This is stricter than `not
is_booted`, i.e. the tests will fail in non-booted non-buildah cases
(and then we can adjust them accordingly, if/when we ever get them).